+++ proxybroker/proxy.py
@@ -60,7 +60,7 @@ class Proxy:
             raise
         return self
 
-    def __init__(self, host=None, port=None, types=(),
+    def __init__(self, host=None, port=None, login=None, password=None, types=(),
                  timeout=8, verify_ssl=False):
         self.host = host
         if not Resolver.host_is_ip(self.host):
@@ -70,6 +70,9 @@ class Proxy:
         self.port = int(port)
         if self.port > 65535:
             raise ValueError('The port of proxy cannot be greater than 65535')
+
+        self.login = login
+        self.password = password
 
         self.expected_types = set(types) & {'HTTP', 'HTTPS', 'CONNECT:80',
                                             'CONNECT:25', 'SOCKS4', 'SOCKS5'}
