+++ proxybroker/utils.py
@@ -25,6 +25,13 @@ IPPortPatternLine = re.compile(
     r'^.*?(?P<ip>(?:(?:25[0-5]|2[0-4]\d|[01]?\d\d?)\.){3}(?:25[0-5]|2[0-4]\d|[01]?\d\d?)).*?(?P<port>\d{2,5}).*$',  # noqa
     flags=re.MULTILINE)
 
+IPPortAuthPatternLine = re.compile(
+    r'^.*?(?P<ip>(?:(?:25[0-5]|2[0-4]\d|[01]?\d\d?)\.){3}(?:25[0-5]|2[0-4]\d|[01]?\d\d?)).*?'
+    r'(?P<port>\d{2,5})'  # noqa
+    r'(?:\s*[:,]*\s*(?P<login>.*?)'
+    r'\s*[:,]\s*(?P<password>.*$))?',
+    flags=re.MULTILINE)
+
 IPPortPatternGlobal = re.compile(
     r'(?P<ip>(?:(?:25[0-5]|2[0-4]\d|[01]?\d\d?)\.){3}(?:25[0-5]|2[0-4]\d|[01]?\d\d?))'  # noqa
     r'(?=.*?(?:(?:(?:(?:25[0-5]|2[0-4]\d|[01]?\d\d?)\.){3}(?:25[0-5]|2[0-4]\d|[01]?\d\d?))|(?P<port>\d{2,5})))',  # noqa
